Updated hideout.json, levels.json and maps.json using https://api.tarkov.dev/
levels : The json format changed from `"1": {"exp": 0, "group": "1"}` to ```"1": {
        "exp": 0,
        "group": "1"
    }```

maps: Added ground-zero but unable to change the "svg" value, please check to make sure it doesnt break anything

hideout: Added missing new modules as well as updated requirements for each module. **I was unable to determine how the id of each module work. So requesting for some info on this so I can set it to the proper number.**